Split out an mc-consensus-scp-types with the core types, for a smaller footprint
