+++ consensus/scp/src/lib.rs
@@ -4,21 +4,26 @@
 #![allow(non_snake_case)]
 #![deny(missing_docs)]
 
-pub mod core_types;
+pub mod ballot;
 pub mod msg;
 pub mod node;
 pub mod predicates;
-pub mod quorum_set;
+pub mod quorum_set_ext;
 pub mod scp_log;
 pub mod slot;
 pub mod slot_state;
+#[cfg(any(test, feature = "test_utils"))]
 pub mod test_utils;
 mod utils;
 
 #[doc(inline)]
+pub use mc_consensus_scp_types::{core_types, quorum_set};
+
+#[doc(inline)]
 pub use crate::{
-    core_types::{CombineFn, GenericNodeId, Identifier, SlotIndex, ValidityFn, Value},
+    core_types::{GenericNodeId, Identifier, SlotIndex, Value},
     msg::{Msg, Topic},
     node::{MockScpNode, Node, ScpNode},
     quorum_set::{QuorumSet, QuorumSetMember},
+    quorum_set_ext::QuorumSetExt,
 };
