+++ consensus/scp/src/slot.rs
@@ -5,31 +5,36 @@
 //! The transactions validated in this slot determine the values to include in
 //! the next block appended to the ledger.
 use crate::{
-    core_types::{Ballot, CombineFn, SlotIndex, ValidityFn, Value},
+    ballot::Ballot,
     msg::*,
     predicates::{
         BallotRangePredicate, BallotSetPredicate, FuncPredicate, Predicate, ValueSetPredicate,
     },
-    quorum_set::QuorumSet,
     slot_state::SlotState,
-    utils,
+    utils, QuorumSet, QuorumSetExt, SlotIndex, Value,
 };
-use core::cmp;
-use maplit::{btreeset, hashset};
 use mc_common::{
     logger::{log, o, Logger},
-    NodeID,
+    HashMap, HashSet, NodeID,
 };
 #[cfg(test)]
 use mockall::*;
 use primitive_types::{U256, U512};
 use serde::{Deserialize, Serialize};
 use std::{
-    collections::{BTreeSet, HashMap, HashSet},
+    cmp,
+    collections::BTreeSet,
     fmt::Display,
     sync::Arc,
     time::{Duration, Instant},
 };
+
+/// Application-specific function for combining multiple values. Must be
+/// deterministic.
+pub type CombineFn<V, E> = Arc<(dyn Fn(&[V]) -> Result<Vec<V>, E> + Sync + Send)>;
+
+/// Application-specific validation of value.
+pub type ValidityFn<V, E> = Arc<(dyn Fn(&V) -> Result<(), E> + Sync + Send)>;
 
 /// The various phases of the SCP protocol.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
@@ -1457,8 +1462,9 @@ impl<V: Value, ValidationError: Display> Slot<V, ValidationError> {
             // Test if a blocking set has issued "accept nominate" for each value.
             for value in candidates {
                 // Test if a blocking set has issued "accept nominate(v)".
+                let values = BTreeSet::from_iter([value.clone()]);
                 let predicate = ValueSetPredicate::<V> {
-                    values: btreeset! {value.clone()},
+                    values,
                     test_fn: Arc::new(|msg, values| match msg.accepts_nominated() {
                         None => BTreeSet::default(),
                         Some(values_accepted_nominated) => values
@@ -1544,8 +1550,9 @@ impl<V: Value, ValidationError: Display> Slot<V, ValidationError> {
             let mut results: HashSet<Ballot<V>> = Default::default();
 
             for ballot in candidates.into_iter() {
+                let ballots = HashSet::from_iter([ballot.clone()]);
                 let predicate = BallotSetPredicate::<V> {
-                    ballots: hashset! { ballot.clone()},
+                    ballots,
                     test_fn: Arc::new(|msg, candidates| {
                         let mut intersections: HashSet<Ballot<V>> = HashSet::default();
 
@@ -1812,8 +1819,8 @@ impl<V: Value, ValidationError: Display> Slot<V, ValidationError> {
 #[cfg(test)]
 mod nominate_protocol_tests {
     use super::*;
-    use crate::{core_types::*, quorum_set::*, test_utils::*};
-    use maplit::{btreeset, hashset};
+    use crate::test_utils::*;
+    use maplit::btreeset;
     use mc_common::logger::test_with_logger;
 
     #[test_with_logger]
@@ -1943,7 +1950,7 @@ mod nominate_protocol_tests {
                 Y: BTreeSet::default(),
             }),
         );
-        slot.X = hashset! { 1234};
+        slot.X = HashSet::from_iter([1234]);
         slot.M.insert(msg_1.sender_id.clone(), msg_1);
         let expected = BTreeSet::default();
         assert_eq!(slot.additional_values_accepted_nominated(), expected);
@@ -2133,8 +2140,8 @@ mod nominate_protocol_tests {
             Arc::new(trivial_combine_fn),
             logger,
         );
-        slot.Y = hashset! { "B"};
-        slot.Z = hashset! { "B"};
+        slot.Y = HashSet::from_iter(["B"]);
+        slot.Z = HashSet::from_iter(["B"]);
         slot.M.insert(msg_2.sender_id.clone(), msg_2);
         slot.M.insert(msg_3.sender_id.clone(), msg_3);
         slot.M.insert(msg_4.sender_id.clone(), msg_4);
@@ -2144,7 +2151,7 @@ mod nominate_protocol_tests {
         // the local node to update it's accepted nominated (Y) list from what
         // the blocking set has agreed on.
         slot.update_YZ();
-        assert_eq!(slot.Y, hashset! { "A", "B", "C", "D"});
+        assert_eq!(slot.Y, HashSet::from_iter(["A", "B", "C", "D"]));
     }
 
     #[test_with_logger]
@@ -2301,8 +2308,8 @@ mod nominate_protocol_tests {
             Arc::new(trivial_combine_fn),
             logger,
         );
-        slot.Y = hashset! { "A", "B", "C", "D"};
-        slot.Z = hashset! { "A", "B", "C"};
+        slot.Y = HashSet::from_iter(["A", "B", "C", "D"]);
+        slot.Z = HashSet::from_iter(["A", "B", "C"]);
         slot.M.insert(msg_1.sender_id.clone(), msg_1);
         slot.M.insert(msg_3.sender_id.clone(), msg_3);
         slot.M.insert(msg_4.sender_id.clone(), msg_4);
@@ -2311,7 +2318,7 @@ mod nominate_protocol_tests {
         // Calling updateYZ should add "D" to confirmed nominated (Z) since a quorum
         // (2,3,4,5) have accepted nominated (Y) it.
         slot.update_YZ();
-        assert_eq!(slot.Z, hashset! { "A", "B", "C", "D"});
+        assert_eq!(slot.Z, HashSet::from_iter(["A", "B", "C", "D"]));
     }
 
     #[test_with_logger]
@@ -2407,8 +2414,8 @@ mod nominate_protocol_tests {
 #[cfg(test)]
 mod ballot_protocol_tests {
     use super::*;
-    use crate::{core_types::*, quorum_set::*, test_utils::*};
-    use maplit::{btreeset, hashset};
+    use crate::test_utils::*;
+    use maplit::btreeset;
     use mc_common::logger::test_with_logger;
 
     // TODO: reject a message if it contains a ballot containing incorrectly ordered
@@ -2690,8 +2697,8 @@ mod ballot_protocol_tests {
         // adding a Prepare message from node 2 as well.
         {
             slot.phase = Phase::NominatePrepare;
-            slot.X = hashset! { 1337, 1338};
-            slot.Y = hashset! { 1234, 5678};
+            slot.X = HashSet::from_iter([1337, 1338]);
+            slot.Y = HashSet::from_iter([1234, 5678]);
             slot.B = Ballot::new(2, &[1234, 5678]);
             slot.P = Some(slot.B.clone());
             slot.last_sent_msg = slot.out_msg();
@@ -2821,8 +2828,8 @@ mod ballot_protocol_tests {
         // Initialize slot so that it has issued "confirm prepare(b)".
         {
             slot.phase = Phase::Prepare;
-            slot.X = hashset! { 1337, 1338};
-            slot.Y = hashset! { 1234, 5678};
+            slot.X = HashSet::from_iter([1337, 1338]);
+            slot.Y = HashSet::from_iter([1234, 5678]);
             slot.B = Ballot::new(3, &[1234, 5678]);
             slot.P = Some(Ballot::new(2, &[1234, 5678]));
             slot.H = slot.P.clone();
@@ -3577,7 +3584,7 @@ mod ballot_protocol_tests {
 
         // Node 1 has issued "vote prepare(b)".
         {
-            slot.Y = hashset! { 1234, 5678};
+            slot.Y = HashSet::from_iter([1234, 5678]);
             slot.B = ballot.clone();
             slot.last_sent_msg = slot.out_msg();
 
@@ -4461,7 +4468,7 @@ mod ballot_protocol_tests {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{core_types::*, test_utils::*};
+    use crate::test_utils::*;
     use mc_common::logger::test_with_logger;
 
     #[test_with_logger]
