+++ consensus/scp/src/slot_state.rs
@@ -5,16 +5,14 @@
 //! might embed a `SlotIndex` directly inside a `Slot`.
 
 use crate::{
-    core_types::{Ballot, SlotIndex, Value},
+    ballot::Ballot,
     msg::*,
     slot::{Phase, Slot},
+    SlotIndex, Value,
 };
-use mc_common::NodeID;
+use mc_common::{HashSet, NodeID};
 use serde::{Deserialize, Serialize};
-use std::{
-    collections::{BTreeSet, HashSet},
-    fmt::Display,
-};
+use std::{collections::BTreeSet, fmt::Display};
 
 /// Serializable slot state used for debugging purposes.
 #[derive(Clone, Serialize, Deserialize)]
