+++ consensus/scp/src/test_utils.rs
@@ -6,13 +6,13 @@
 // code. https://github.com/rust-lang/rust/issues/46379
 #![allow(dead_code)]
 
-use crate::{core_types::Value, slot::Slot, QuorumSet, SlotIndex};
-use mc_common::{logger::Logger, NodeID, ResponderId};
-use mc_crypto_keys::Ed25519Pair;
-use mc_util_from_random::FromRandom;
-use rand::SeedableRng;
-use rand_hc::Hc128Rng as FixedRng;
-use std::{fmt, str::FromStr, sync::Arc};
+pub use mc_consensus_scp_types::test_utils::{
+    fig_2_network, test_node_id, test_node_id_and_signer, three_node_cycle, three_node_dense_graph,
+};
+
+use crate::{slot::Slot, QuorumSet, SlotIndex, Value};
+use mc_common::{logger::Logger, NodeID};
+use std::{fmt, sync::Arc};
 
 /// Error for transaction validation
 #[derive(Clone)]
@@ -49,29 +49,6 @@ pub fn get_bounded_combine_fn<V: Value>(
     }
 }
 
-/// Creates NodeID from integer for testing.
-pub fn test_node_id(node_id: u32) -> NodeID {
-    let (node_id, _signer) = test_node_id_and_signer(node_id);
-    node_id
-}
-
-/// Creates NodeID and Signer keypair from integer for testing.
-pub fn test_node_id_and_signer(node_id: u32) -> (NodeID, Ed25519Pair) {
-    let mut seed_bytes = [0u8; 32];
-    let node_id_bytes = node_id.to_be_bytes();
-    seed_bytes[..node_id_bytes.len()].copy_from_slice(&node_id_bytes[..]);
-
-    let mut seeded_rng: FixedRng = SeedableRng::from_seed(seed_bytes);
-    let signer_keypair = Ed25519Pair::from_random(&mut seeded_rng);
-    (
-        NodeID {
-            responder_id: ResponderId::from_str(&format!("node{}.test.com:8443", node_id)).unwrap(),
-            public_key: signer_keypair.public_key(),
-        },
-        signer_keypair,
-    )
-}
-
 /// Creates a new slot.
 pub fn get_slot(
     slot_index: SlotIndex,
@@ -87,84 +64,4 @@ pub fn get_slot(
         Arc::new(trivial_combine_fn),
         logger,
     )
-}
-
-/// Three nodes that form a three-node cycle.
-///
-/// * Node 1 has the quorum slice {1,2}, where {2} is a blocking set.
-/// * Node 2 has the quorum slice {2,3}, where {3} is a blocking set.
-/// * Node 3 has the quorum slice {1,3}, where {1} is a blocking set.
-/// * The only quorum is the set of all three nodes {1, 2, 3}.
-pub fn three_node_cycle() -> (
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-) {
-    let node_1 = (
-        test_node_id(1),
-        QuorumSet::new_with_node_ids(1, vec![test_node_id(2)]),
-    );
-    let node_2 = (
-        test_node_id(2),
-        QuorumSet::new_with_node_ids(1, vec![test_node_id(3)]),
-    );
-    let node_3 = (
-        test_node_id(3),
-        QuorumSet::new_with_node_ids(1, vec![test_node_id(1)]),
-    );
-    (node_1, node_2, node_3)
-}
-
-/// The four-node network from Fig. 2 of the [Stellar Whitepaper](https://www.stellar.org/papers/stellar-consensus-protocol).
-///
-/// * Node 1 has the quorum slice {1,2,3}, where {2}, {3}, {2,3} are blocking
-///   sets.
-/// * Nodes 2,3, and 4 have the quorum slice {2,3,4}.
-/// * The only quorum is the set of all nodes {1,2,3,4}.
-pub fn fig_2_network() -> (
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-) {
-    let node_1 = (
-        test_node_id(1),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(2), test_node_id(3)]),
-    );
-    let node_2 = (
-        test_node_id(2),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(3), test_node_id(4)]),
-    );
-    let node_3 = (
-        test_node_id(3),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(2), test_node_id(4)]),
-    );
-    let node_4 = (
-        test_node_id(4),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(2), test_node_id(4)]),
-    );
-
-    (node_1, node_2, node_3, node_4)
-}
-
-/// A three-node network where the only quorum is the set of all three nodes.
-/// Each node is a blocking set for each other.
-pub fn three_node_dense_graph() -> (
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-    (NodeID, QuorumSet),
-) {
-    let node_1 = (
-        test_node_id(1),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(2), test_node_id(3)]),
-    );
-    let node_2 = (
-        test_node_id(2),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(1), test_node_id(3)]),
-    );
-    let node_3 = (
-        test_node_id(3),
-        QuorumSet::new_with_node_ids(2, vec![test_node_id(1), test_node_id(2)]),
-    );
-    (node_1, node_2, node_3)
 }
