+++ consensus/scp/types/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "mc-consensus-scp-types"
+version = "1.3.0-pre0"
+authors = ["MobileCoin"]
+edition = "2018"
+readme = "../README.md"
+
+[features]
+default = []
+std = ["mc-common/std", "serde/std"]
+test_utils = ["mc-util-test-helper", "rand", "rand_hc"]
+
+[dependencies]
+mc-common = { path = "../../../common", default-features = false }
+mc-crypto-digestible = { path = "../../../crypto/digestible", features = ["derive"] }
+mc-crypto-keys = { path = "../../../crypto/keys" }
+mc-util-from-random = { path = "../../../util/from-random" }
+mc-util-test-helper = { path = "../../../util/test-helper", optional = true }
+
+rand = { version = "0.8", default-features = false, optional = true }
+rand_hc = { version = "0.3", default-features = false, optional = true }
+serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+
+[dev-dependencies]
+mc-util-serial = { path = "../../../util/serial", features = ["std"] }
+mc-util-test-helper = { path = "../../../util/test-helper" }
+
+rand = "0.8"
+rand_hc = "0.3"
